+++ AgroFamily/ViewModel/BusinessStatusViewModel.cs
@@ -29,7 +29,7 @@ namespace AgroFamily.ViewModel
 
             ISaleRepository saleRepository = new SaleRepository();
             IExpensiveModel expensiveRepository = new ExpensiveRepository();
-            int AmountSales_int = saleRepository.GetAmountInAMonth(month, year);
+            long AmountSales_int = saleRepository.GetAmountInAMonth(month, year);
             double AmountExp_int = expensiveRepository.GetAmountInAMonth(month, year);
             double AmountDiff_int = (AmountSales_int - AmountExp_int);
